# docs/api/ssvnode.openapi.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP endpoints exposed by the SSV node, including exporter traces and validator metadata.",
        "title": "SSV Node API",
        "contact": {
            "name": "SSV Labs",
            "url": "https://ssvlabs.io",
            "email": "info@ssv.network"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/exporter/decideds": {
            "get": {
                "description": "Returns decided duty participant traces for validators or committees, including partial error details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve decided message traces",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "From is the starting slot (inclusive).",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Indices is a comma-separated list of validator indices.",
                        "name": "indices",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "PubKeys is a comma-separated list of validator public keys (hex, 96 chars per key).",
                        "name": "pubkeys",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "ATTESTER",
                                "AGGREGATOR",
                                "PROPOSER",
                                "SYNC_COMMITTEE",
                                "SYNC_COMMITTEE_CONTRIBUTION"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Roles is a comma-separated list of beacon roles to include.",
                        "name": "roles",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "To is the ending slot (inclusive).",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.DecidedsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.TraceDecidedsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns decided duty participant traces for validators or committees, including partial error details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve decided message traces",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "From is the starting slot (inclusive).",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Indices is a comma-separated list of validator indices.",
                        "name": "indices",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "PubKeys is a comma-separated list of validator public keys (hex, 96 chars per key).",
                        "name": "pubkeys",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "ATTESTER",
                                "AGGREGATOR",
                                "PROPOSER",
                                "SYNC_COMMITTEE",
                                "SYNC_COMMITTEE_CONTRIBUTION"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Roles is a comma-separated list of beacon roles to include.",
                        "name": "roles",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "To is the ending slot (inclusive).",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.DecidedsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.TraceDecidedsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/exporter/traces/committee": {
            "get": {
                "description": "Returns consensus and post-consensus traces for requested committees.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve committee duty traces",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "CommitteeIDs is a comma-separated list of committee IDs (hex, 64 chars per ID).",
                        "name": "committeeIDs",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "From is the starting slot (inclusive).",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "To is the ending slot (inclusive).",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.CommitteeTracesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.CommitteeTracesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns consensus and post-consensus traces for requested committees.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve committee duty traces",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "CommitteeIDs is a comma-separated list of committee IDs (hex, 64 chars per ID).",
                        "name": "committeeIDs",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "From is the starting slot (inclusive).",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "To is the ending slot (inclusive).",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.CommitteeTracesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.CommitteeTracesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/exporter/auditor/findings": {
            "get": {
                "description": "Returns persisted trace<->schedule mismatch findings (if auditor is enabled).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve auditor findings",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "From is the starting slot (inclusive).",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "description": "To is the ending slot (inclusive).",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Optional reason code filter.",
                        "name": "reason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Optional role filter (attester or sync_committee).",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Optional SSV committee ID (hex, 64 chars).",
                        "name": "committeeID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Optional validator index filter.",
                        "name": "validatorIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Maximum number of findings to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.AuditorFindingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.AuditorFindingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Auditor disabled",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Returns persisted trace<->schedule mismatch findings (if auditor is enabled).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporter"
                ],
                "summary": "Retrieve auditor findings",
                "parameters": [
                    {
                        "description": "Filters as JSON body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/exporter.AuditorFindingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/exporter.AuditorFindingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Auditor disabled",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
...
        "exporter.CommitteeSchedule": {
            "type": "object",
            "properties": {
                "committeeID": {
                    "type": "string",
                    "format": "hex"
                },
                "roles": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                },
                "slot": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "exporter.CommitteeTracesRequest": {
            "type": "object",
            "properties": {
                "committeeIDs": {
                    "description": "CommitteeIDs is a comma-separated list of committee IDs (hex, 64 chars per ID).",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "hex",
                        "maxLength": 64,
                        "minLength": 64
                    }
                },
                "from": {
                    "description": "From is the starting slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                },
                "to": {
                    "description": "To is the ending slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                }
            }
        },
        "exporter.CommitteeTracesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data contains the list of committee duty traces matching the request.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exporter.CommitteeTrace"
                    }
                },
                "errors": {
                    "description": "Errors lists non-fatal issues encountered while building the response (duties not found, enrichment errors, etc.).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "committee duty missing for slot 123456"
                    ]
                },
                "schedule": {
                    "description": "Schedule lists requested duties unioned at the committee-level by role.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exporter.CommitteeSchedule"
                    }
                }
            }
        },
        "exporter.AuditorFindingsRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "From is the starting slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                },
                "to": {
                    "description": "To is the ending slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                },
                "reason": {
                    "description": "Optional reason code filter.",
                    "type": "string"
                },
                "role": {
                    "description": "Optional role filter (attester or sync_committee).",
                    "type": "string"
                },
                "committeeID": {
                    "description": "Optional SSV committee ID (hex, 64 chars).",
                    "type": "string",
                    "format": "hex"
                },
                "validatorIndex": {
                    "description": "Optional validator index filter.",
                    "type": "integer",
                    "format": "int64"
                },
                "limit": {
                    "description": "Maximum number of findings to return.",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "exporter.AuditorFindingsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auditor.Finding"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "exporter.DecidedParticipant": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "object",
                    "properties": {
                        "Signers": {
                            "description": "We're keeping \"Signers\" capitalized to avoid breaking existing clients that rely on the current structure.",
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "public_key": {
                    "type": "string",
                    "format": "hex"
                },
                "role": {
                    "type": "string",
                    "example": "ATTESTER"
                },
                "slot": {
                    "type": "integer",
                    "format": "int64",
                    "example": 123456
                }
            }
        },
        "exporter.DecidedsRequest": {
            "type": "object",
            "required": [
                "roles"
            ],
            "properties": {
                "from": {
                    "description": "From is the starting slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                },
                "indices": {
                    "description": "Indices is a comma-separated list of validator indices.",
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    }
                },
                "pubkeys": {
                    "description": "PubKeys is a comma-separated list of validator public keys (hex, 96 chars per key).",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "hex",
                        "maxLength": 96,
                        "minLength": 96
                    }
                },
                "roles": {
                    "description": "Roles is a comma-separated list of beacon roles to include.",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "ATTESTER",
                            "AGGREGATOR",
                            "PROPOSER",
                            "SYNC_COMMITTEE",
                            "SYNC_COMMITTEE_CONTRIBUTION"
                        ]
                    }
                },
                "to": {
                    "description": "To is the ending slot (inclusive).",
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                }
            }
        },
...
        "exporter.TraceDecidedsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data contains the decided duty participant entries matching the request.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exporter.DecidedParticipant"
                    }
                },
                "errors": {
                    "description": "Errors lists non-fatal issues encountered while building the response (e.g., entries with empty signer sets).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "omitting entry with no signers (index=deadbeef",
                        " slot=123456",
                        " role=ATTESTER)"
                    ]
                }
            }
        },
...
    },
...
}